Rename check-input-format to fingerprinting
