+++ libs/hdf-converters/index.ts
@@ -26,6 +26,6 @@ export * from './src/sonarqube-mapper';
 export * from './src/splunk-mapper';
 export * from './src/twistlock-mapper';
 export * from './src/utils/attestations';
-export * from './src/utils/check-input-format';
+export * from './src/utils/fingerprinting';
 export * from './src/xccdf-results-mapper';
 export * from './src/zap-mapper';
