+++ libs/hdf-converters/src/utils/fingerprinting.ts
